minimizde code redundancy and duplicity
