+++ C/Linked List/doublyLinkedList.c
@@ -1,287 +1,263 @@
-#include<stdio.h>
-#include<stdlib.h>
-
-struct node {
-    struct node * prev;
-    struct node * next;
+/*
+program to implement following operations on the doubly linked list.
+a. Insert a node at the front of the linked list.
+b. Insert a node at the end of the linked list. 
+c. Delete a last node of the linked list.
+d. Delete a node before specified position/
+*/
+#include <stdio.h>
+#include <conio.h>
+#include <malloc.h>
+struct node
+{
+    struct node *next;
     int data;
+    struct node *prev;
 };
-
-// All Basic Operations of Doubly Linked List
-struct node * head;
-void insertAtStart();
-void insertAtEnd();
-void insertAtPos();
-void deleteAtStart();
-void deleteAtEnd();
-void deleteAtPos();
-void print();
-void search();
-
-void main() {
-    int choice = 0;
-    while (choice != 13) {
-        printf("1.Insert At Start   | 2.Insert At Last   | 3.Insert At Position\n4.Delete From Start | 5.Delete From last | 6.Delete From Postition\n7.Find\n8.Length\n9.Middle\n10.Reverse & Show\n11.Show\n12.Exit\n");
-        printf("\nEnter Choice: ");
-        scanf("\n%d", & choice);
-        switch (choice) {
+struct node *start = NULL;
+struct node *create_ll(struct node *);
+struct node *display(struct node *);
+struct node *insert_beg(struct node *);
+struct node *insert_end(struct node *);
+struct node *insert_before(struct node *);
+struct node *insert_after(struct node *);
+struct node *delete_beg(struct node *);
+struct node *delete_end(struct node *);
+struct node *delete_before(struct node *);
+struct node *delete_after(struct node *);
+struct node *delete_list(struct node *);
+int main()
+{
+    int option;
+    clrscr();
+    do
+    {
+        printf("\n\n *****MAIN MENU *****");
+        printf("\n 1: Create a list");
+        printf("\n 2: Display the list");
+        printf("\n 3: Add a node at the beginning");
+        printf("\n 4: Add a node at the end");
+        printf("\n 5: Add a node before a given node");
+        printf("\n 6: Add a node after a given node");
+        printf("\n 7: Delete a node from the beginning");
+        printf("\n 8: Delete a node from the end");
+        printf("\n 9: Delete a node before a given node");
+        printf("\n 10: Delete a node after a given node");
+        printf("\n 11: Delete the entire list");
+        printf("\n 12: EXIT");
+        printf("\n\n Enter your option : ");
+        scanf("%d", &option);
+        switch (option)
+        {
         case 1:
-            insertAtStart();
+            start = create_ll(start);
+            printf("\n DOUBLY LINKED LIST CREATED");
             break;
         case 2:
-            insertAtEnd();
+            start = display(start);
             break;
         case 3:
-            insertAtPos();
+            start = insert_beg(start);
             break;
         case 4:
-            deleteAtStart();
+            start = insert_end(start);
             break;
         case 5:
-            deleteAtEnd();
+            start = insert_before(start);
             break;
         case 6:
-            deleteAtPos();
+            start = insert_after(start);
             break;
         case 7:
-            search();
+            start = delete_beg(start);
             break;
         case 8:
-            size();
+            start = delete_end(start);
             break;
         case 9:
-            middle(head);
+            start = delete_before(start);
             break;
         case 10:
-            reverse( & head);
-            print();
+            start = delete_after(start);
             break;
         case 11:
-            print();
+            start = delete_list(start);
+            printf("\n DOUBLY LINKED LIST DELETED");
             break;
-        case 12:
-            exit(0);
-            break;
-        default:
-            printf("Please enter valid choice..");
         }
-    }
+    } while (option != 12);
+    getch();
+    return 0;
 }
-
-// Respective Function Codes
-void insertAtStart() {
-    struct node * ptr;
-    int item;
-    ptr = (struct node * ) malloc(sizeof(struct node));
-    if (ptr == NULL) {
-        printf("OVERFLOW");
-    } else {
-        printf("Enter Item value: ");
-        scanf("%d", & item);
-
-        if (head == NULL) {
-            ptr -> next = NULL;
-            ptr -> prev = NULL;
-            ptr -> data = item;
-            head = ptr;
-        } else {
-            ptr -> data = item;
-            ptr -> prev = NULL;
-            ptr -> next = head;
-            head -> prev = ptr;
-            head = ptr;
+struct node *create_ll(struct node *start)
+{
+    struct node *new_node, *ptr;
+    int num;
+    printf("\n Enter –1 to end");
+    printf("\n Enter the data : ");
+    scanf("%d", &num);
+    while ( num != -1 )
+    {
+        if (start == NULL)
+        {
+            new_node = (struct node *)malloc(sizeof(struct node));
+            new_node->prev = NULL;
+            new_node->data = num;
+            new_node->next = NULL;
+            start = new_node;
         }
-        printf("Node Inserted\n\n");
-    }
-}
-
-void insertAtEnd() {
-    struct node * ptr, * temp;
-    int item;
-    ptr = (struct node * ) malloc(sizeof(struct node));
-    if (ptr == NULL) {
-        printf("OVERFLOW");
-    } else {
-        printf("Enter value: ");
-        scanf("%d", & item);
-        ptr -> data = item;
-        if (head == NULL) {
-            ptr -> next = NULL;
-            ptr -> prev = NULL;
-            head = ptr;
-        } else {
-            temp = head;
-            while (temp -> next != NULL) {
-                temp = temp -> next;
-            }
-            temp -> next = ptr;
-            ptr -> prev = temp;
-            ptr -> next = NULL;
+        else
+        {
+            ptr = start;
+            new_node = (struct node *)malloc(sizeof(struct node));
+            new_node-> data = num;
+            while (ptr-> next != NULL)
+                ptr = ptr-> next;
+            ptr-> next = new_node;
+            new_node-> prev = ptr;
+            new_node-> next = NULL;
         }
-
+        printf("\n Enter the data : ");
+        scanf("%d", &num);
     }
-    printf("Node Inserted\n\n");
+    return start;
 }
-
-void insertAtPos() {
-    struct node * ptr, * temp;
-    int item, loc, i;
-    ptr = (struct node * ) malloc(sizeof(struct node));
-    if (ptr == NULL) {
-        printf("OVERFLOW");
-    } else {
-        temp = head;
-        printf("Enter the location: ");
-        scanf("%d", & loc);
-        for (i = 0; i < loc; i++) {
-            temp = temp -> next;
-            if (temp == NULL) {
-                printf("There are less than %d elements", loc);
-                return;
-            }
-        }
-        printf("Enter value: ");
-        scanf("%d", & item);
-        ptr -> data = item;
-        ptr -> next = temp -> next;
-        ptr -> prev = temp;
-        temp -> next = ptr;
-        temp -> next -> prev = ptr;
-        printf("Node Inserted\n\n");
+struct node *display(struct node *start)
+{
+    struct node *ptr;
+    ptr = start;
+    while (ptr != NULL)
+    {
+        printf("\t %d", ptr->data);
+        ptr = ptr->next;
     }
+    return start;
 }
-
-void deleteAtStart() {
-    struct node * ptr;
-    if (head == NULL) {
-        printf("\n UNDERFLOW");
-    } else if (head -> next == NULL) {
-        head = NULL;
-        free(head);
-        printf("\nnode deleted\n");
-    } else {
-        ptr = head;
-        head = head -> next;
-        head -> prev = NULL;
-        free(ptr);
-        printf("\nnode deleted\n");
-    }
-
+struct node *insert_beg(struct node *start)
+{
+    struct node *new_node;
+    int num;
+    printf("\n Enter the data : ");
+    scanf("%d", &num);
+    new_node = (struct node *)malloc(sizeof(struct node));
+    new_node->data = num;
+    start->prev = new_node;
+    new_node->next = start;
+    new_node->prev = NULL;
+    start = new_node;
+    return start;
 }
-
-void deleteAtEnd() {
-    struct node * ptr;
-    if (head == NULL) {
-        printf("\n UNDERFLOW");
-    } else if (head -> next == NULL) {
-        head = NULL;
-        free(head);
-        printf("\nnode deleted\n");
-    } else {
-        ptr = head;
-        if (ptr -> next != NULL) {
-            ptr = ptr -> next;
-        }
-        ptr -> prev -> next = NULL;
-        free(ptr);
-        printf("\nnode deleted\n");
-    }
+struct node *insert_end(struct node *start)
+{
+    struct node *ptr, *new_node;
+    int num;
+    printf("\n Enter the data : ");
+    scanf("%d", &num);
+    new_node = (struct node *)malloc(sizeof(struct node));
+    new_node->data = num;
+    ptr = start;
+    while (ptr->next != NULL)
+        ptr = ptr->next;
+    ptr->next = new_node;
+    new_node->prev = ptr;
+    new_node->next = NULL;
+    return start;
 }
-
-void deleteAtPos() {
-    struct node * ptr, * temp;
+struct node *insert_before(struct node *start)
+{
+    struct node *new_node, *ptr;
+    int num, val;
+    printf("\n Enter the data : ");
+    scanf("%d", &num);
+    printf("\n Enter the value before which the data has to be inserted:");
+    scanf("%d", &val);
+    new_node = (struct node *)malloc(sizeof(struct node));
+    new_node->data = num;
+    ptr = start;
+    while (ptr->data != val)
+        ptr = ptr->next;
+    new_node->next = ptr;
+    new_node->prev = ptr->prev;
+    ptr->prev->next = new_node;
+    ptr->prev = new_node;
+    return start;
+}
+struct node *insert_after(struct node *start)
+{
+    struct node *new_node, *ptr;
+    int num, val;
+    printf("\n Enter the data : ");
+    scanf("%d", &num);
+    printf("\n Enter the value after which the data has to be inserted:");
+    scanf("%d", &val);
+    new_node = (struct node *)malloc(sizeof(struct node));
+    new_node->data = num;
+    ptr = start;
+    while (ptr->data != val)
+        ptr = ptr->next;
+    new_node->prev = ptr;
+    new_node->next = ptr->next;
+    ptr->next->prev = new_node;
+    ptr->next = new_node;
+    return start;
+}
+struct node *delete_beg(struct node *start)
+{
+    struct node *ptr;
+    ptr = start;
+    start = start->next;
+    start->prev = NULL;
+    free(ptr);
+    return start;
+}
+struct node *delete_end(struct node *start)
+{
+    struct node *ptr;
+    ptr = start;
+    while (ptr->next != NULL)
+        ptr = ptr->next;
+    ptr->prev->next = NULL;
+    free(ptr);
+    return start;
+}
+struct node *delete_after(struct node *start)
+{
+    struct node *ptr, *temp;
     int val;
-    printf("\n Enter the data after which the node is to be deleted : ");
-    scanf("%d", & val);
-    ptr = head;
-    while (ptr -> data != val)
-        ptr = ptr -> next;
-    if (ptr -> next == NULL) {
-        printf("\nCan't delete\n");
-    } else if (ptr -> next -> next == NULL) {
-        ptr -> next = NULL;
-    } else {
-        temp = ptr -> next;
-        ptr -> next = temp -> next;
-        temp -> next -> prev = ptr;
-        free(temp);
-        printf("\nnode deleted\n");
-    }
+    printf("\n Enter the value after which the node has to deleted : ");
+    scanf("%d", &val);
+    ptr = start;
+    while (ptr->data != val)
+        ptr = ptr->next;
+    temp = ptr->next;
+    ptr->next = temp->next;
+    temp->next->prev = ptr;
+    free(temp);
+    return start;
 }
-
-void search() {
-    struct node * ptr;
-    int item, i = 0, flag;
-    ptr = head;
-    if (ptr == NULL) {
-        printf("\nEmpty List\n");
-    } else {
-        printf("\nEnter item which you want to search?\n");
-        scanf("%d", & item);
-        while (ptr != NULL) {
-            if (ptr -> data == item) {
-                printf("\nitem found at location %d ", i + 1);
-                flag = 0;
-                break;
-            } else {
-                flag = 1;
-            }
-            i++;
-            ptr = ptr -> next;
-        }
-        if (flag == 1) {
-            printf("\nItem not found\n");
-        }
+struct node *delete_before(struct node *start)
+{
+    struct node *ptr, *temp;
+    int val;
+    printf("\n Enter the value before which the node has to deleted:");
+    scanf("%d", &val);
+    ptr = start;
+    while (ptr->data != val)
+        ptr = ptr->next;
+    temp = ptr->prev;
+    if (temp == start)
+        start = delete_beg(start);
+    else
+    {
+        ptr->prev = temp->prev;
+        temp->prev->next = ptr;
     }
+    free(temp);
+    return start;
 }
-
-void size() {
-    int size = 0;
-    struct node * ptr;
-    if (head == NULL) {
-        printf("\n Size of list is %d\n", size);
-        return;
-    }
-    ptr = head;
-    while (ptr != NULL) {
-        ptr = ptr -> next;
-        size++;
-    }
-    printf("\n Size of list is %d\n", size);
-}
-
-void middle(struct node * head) {
-    struct node * slow_ptr = head;
-    struct node * fast_ptr = head;
-    if (head != NULL) {
-        while (fast_ptr != NULL && fast_ptr -> next != NULL) {
-            fast_ptr = fast_ptr -> next -> next;
-            slow_ptr = slow_ptr -> next;
-        }
-        printf("The middle element is [%d]\n\n", slow_ptr -> data);
-    }
-}
-
-void reverse(struct node ** head_ref) {
-    struct node * temp = NULL;
-    struct node * current = * head_ref;
-    while (current != NULL) {
-        temp = current -> prev;
-        current -> prev = current -> next;
-        current -> next = temp;
-        current = current -> prev;
-    }
-    if (temp != NULL)
-        *
-        head_ref = temp -> prev;
-}
-
-void print() {
-    struct node * ptr;
-    printf("\nPrinting values: \n");
-    ptr = head;
-    while (ptr != NULL) {
-        printf("%d\n", ptr -> data);
-        ptr = ptr -> next;
-    }
+struct node *delete_list(struct node *start)
+{
+    while (start != NULL)
+        start = delete_beg(start);
+    return start;
 }
